requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
# coverage==7.3.2
cycler==0.11.0
distlib==0.3.7
Django==4.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.2
fonttools==4.39.3
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.5
MarkupSafe==2.1.3
# matplotlib==3.7.1
# numpy==1.24.2
packaging==23.0
Pillow==9.5.0
platformdirs==3.9.1
psycopg2-binary==2.9.9
pygal==3.0.0
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
# requests==2.31.0
# scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.24.0
whitenoise==6.5.0
